+++ src/edie/evaluator.py
@@ -25,20 +25,24 @@ class Edie(object):
         self.aggregate_evaluators: []
 
         self.report = {"endpoint": api_client.endpoint, "available": True, "dictionaries": {}}
-        self.entries_offset = 0
         self.entries_limit = 100
 
-    def load_dictionaries(self, dictionaries: [str] = None):
+    def load_dictionaries(self, dictionaries: [str] = None, limit=-1):
         dictionary_ids = dictionaries if dictionaries is not None else self.lexonomy_client.dictionaries()["dictionaries"]
-        sys.stderr.write("Evaluating %d dictionaries\n" % len(dictionary_ids))
+        if limit == -1:
+            limit = len(dictionary_ids)
+        sys.stderr.write(f'Evaluating {len(dictionary_ids):d} dictionaries\n')
+        count = 0
         for dictionary_id in dictionary_ids:
             try:
-                sys.stderr.write("Loading Metadata of %s \n" % dictionary_id)
-                metadata = Metadata(self.lexonomy_client.about(dictionary_id))
-                dictionary = Dictionary(dictionary_id, metadata)
-                self.dictionaries.append(dictionary)
+                if count < limit:
+                    sys.stderr.write(f"Loading Metadata of {dictionary_id} \n")
+                    metadata = Metadata(self.lexonomy_client.about(dictionary_id))
+                    dictionary = Dictionary(dictionary_id, metadata)
+                    self.dictionaries.append(dictionary)
+                count+=1
             except HTTPError:
-                sys.stderr.write("Failed loading %s dictionary \n" % dictionary_id)
+                sys.stderr.write(f'Failed loading {dictionary_id} dictionary \n')
 
         return self.dictionaries
 
@@ -46,7 +50,7 @@ class Edie(object):
         for dictionary in self.dictionaries:
             self._prepare_report(dictionary)
 
-            sys.stderr.write("Evaluating %s" % dictionary)
+            sys.stderr.write(f'Evaluating {dictionary}')
 
             metadata_report = {}
             metadata = dictionary.metadata
@@ -64,56 +68,13 @@ class Edie(object):
             self._prepare_report(dictionary)
             entry_report = {}
 
-            max_entries = max_entries if max_entries is not None else dictionary.metadata.entry_count
-            self._loop_entries_endpoint(dictionary, entry_report, max_entries)
+            entry_counter = max_entries if max_entries is not None else dictionary.metadata.entry_count
+            self._loop_entries_endpoint(dictionary, entry_report, entry_counter)
 
             sys.stderr.write("\n")
 
             self._collect_entry_metrics(entry_report)
             self._add_entry_report(dictionary, entry_report)
-
-    def _loop_entries_endpoint(self, dictionary, entry_report, max_entries):
-        while self.entries_offset <= max_entries:
-            try:
-                entries = self.lexonomy_client.list(dictionary.id, limit=self.entries_limit, offset=self.entries_offset)
-                if not entries:
-                    break
-                self._handle_entries(dictionary, entries, entry_report, max_entries)
-                sys.stderr.write(".")
-                sys.stderr.flush()
-                if len(entries) < self.entries_limit:
-                    break
-
-            except HTTPError:
-                self._add_errors(entry_report, f'Failed to retrieve lemmas for dictionary {dictionary.id}')
-            except ParseError as pe:
-                self._add_errors(entry_report, str(pe))
-            except JSONDecodeError as jde:
-                self._add_errors(entry_report, str(jde))
-
-    def _collect_entry_metrics(self, entry_report):
-        for entry_metric in self.entry_metrics_evaluators:
-            if entry_metric.result():
-                sys.stderr.write(str(entry_metric))
-                sys.stderr.write(str(entry_metric.result()))
-                sys.stderr.write("\n")
-                sys.stderr.flush()
-                entry_report.update(entry_metric.result())
-            entry_metric.reset()
-
-    def _handle_entries(self, dictionary, entries, entry_report, max_entries):
-            for entry in entries:
-                self.entries_offset += 1
-                if self.entries_offset > max_entries:
-                    break
-                entry = Entry(entry)
-                if entry.errors:
-                    self._add_errors(entry_report, entry.errors)
-                else:
-                    self._entry_report(dictionary.id, entry_report, entry)
-
-    def evaluation_report(self):
-        return self.report
 
     def entry_evaluation_report_as_dataframe(self):
         return pd.DataFrame.from_dict({i: self.report['dictionaries'][i]['entry_report']
@@ -129,11 +90,70 @@ class Edie(object):
         return self.report['dictionaries'][dictionary_id]['entry_report']
 
     def visualize(self):
-        df = self.entry_evaluation_report_as_dataframe().drop('errors', axis=1)
-        df = df.apply(lambda x: x / x.max(), axis=0)
-        df['dict_type'] = 0
-        parallel_coordinates(df, "dict_type", axvlines=True)
+        dataframe = self.entry_evaluation_report_as_dataframe().drop('errors', axis=1)
+        dataframe = dataframe.apply(lambda x: x / x.max(), axis=0)
+        dataframe['dict_type'] = 0
+        parallel_coordinates(dataframe, "dict_type", axvlines=True)
         plt.show()
+
+    def aggregated_evaluation(self):
+        df = self.metadata_evaluation_report_as_dataframe()
+
+        self.report[AGGREGATION_METRICS] = {
+            DICTIONARY_SIZE: {
+                'min': df[SIZE_OF_DICTIONARY].min(),
+                'max': df[SIZE_OF_DICTIONARY].max(),
+                'mean': df[SIZE_OF_DICTIONARY].mean(),
+                'median': df[SIZE_OF_DICTIONARY].median()
+            }
+        }
+
+    def _loop_entries_endpoint(self, dictionary, entry_report, max_entries):
+        entries_offset = 0
+        while entries_offset <= max_entries:
+            entries = self.lexonomy_client.list(dictionary.id, limit=self.entries_limit, offset=entries_offset)
+            if not entries:
+                break
+            entries_offset = self._handle_entries(dictionary, entries, entry_report, max_entries, entries_offset)
+            sys.stderr.write(str(entries_offset) + '...')
+            sys.stderr.flush()
+            if len(entries) < self.entries_limit:
+                break
+
+    def _collect_entry_metrics(self, entry_report):
+        for entry_metric in self.entry_metrics_evaluators:
+            if entry_metric.result():
+                sys.stderr.write(str(entry_metric))
+                sys.stderr.write(str(entry_metric.result()))
+                sys.stderr.write("\n")
+                sys.stderr.flush()
+                entry_report.update(entry_metric.result())
+            entry_metric.reset()
+
+    def _handle_entries(self, dictionary, entries, entry_report, max_entries, entries_offset):
+        for entry in entries:
+            entries_offset += 1
+            print(entries_offset)
+            if entries_offset > max_entries:
+                break
+            try:
+                entry = Entry(entry)
+                if entry.errors:
+                    self._add_errors(entry_report, entry.errors)
+                else:
+                    self._entry_report(dictionary.id, entry_report, entry)
+
+            except HTTPError:
+                self._add_errors(entry_report, f'Failed to retrieve lemmas for dictionary {dictionary.id}')
+            except ParseError as parse_error:
+                self._add_errors(entry_report, [str(parse_error)])
+            except JSONDecodeError as json_decode_error:
+                self._add_errors(entry_report, [str(json_decode_error)])
+
+        return entries_offset
+
+    def evaluation_report(self):
+        return self.report
 
     def _entry_report(self, dictionary_id: str, entry_report: dict, entry: Entry):
         retrieved_entry: JsonEntry = self._retrieve_entry(dictionary_id, entry)
@@ -149,7 +169,7 @@ class Edie(object):
             try:
                 return JsonEntry(self.lexonomy_client.json(dictionary_id, entry.id))
             except JSONDecodeError as jde:
-                raise JSONDecodeError("Error parsing json response %s: %s" % (entry.id, str(jde)))
+                raise JSONDecodeError(f"Error parsing json response {entry.id}: {str(jde)}")
 
         elif "tei" in entry.formats:
             tei_entry = self.lexonomy_client.tei(dictionary_id, entry.id)
@@ -157,7 +177,7 @@ class Edie(object):
                 tei_entry_element = validate_tei(tei_entry)
                 return JsonEntry.from_tei_entry(tei_entry_element, entry.id)
             except ParseError as pe:
-                raise ParseError("Error with entry %s: %s" % (entry.id, str(pe)))
+                raise ParseError(f"Error with entry {entry.id}: {str(pe)}")
 
     def _run_entry_metrics_evaluators(self, entry_details, entry_metadata):
         for entry_metric in self.entry_metrics_evaluators:
@@ -170,19 +190,8 @@ class Edie(object):
     def _add_entry_report(self, dictionary, entry_report):
         self.report['dictionaries'][dictionary.id]['entry_report'] = entry_report
 
-    def _add_errors(self, entry_report, errors):
+    @staticmethod
+    def _add_errors(entry_report, errors):
         if "errors" not in entry_report:
             entry_report["errors"] = []
         entry_report["errors"].extend(errors)
-
-    def aggregated_evaluation(self):
-        df = self.metadata_evaluation_report_as_dataframe()
-
-        self.report[AGGREGATION_METRICS] = {
-            DICTIONARY_SIZE: {
-                'min': df[SIZE_OF_DICTIONARY].min(),
-                'max': df[SIZE_OF_DICTIONARY].max(),
-                'mean': df[SIZE_OF_DICTIONARY].mean(),
-                'median': df[SIZE_OF_DICTIONARY].median()
-            }
-        }
