+++ test/test_entry.py
@@ -1,7 +1,11 @@
 """
     ELEXIS Protocol for accessing dictionaries
 
-    This protocol allows data to be shared with the ELEXIS platform and should be implemented by all providers of data to the ELEXIS platform. This is an OpenAPI documentation, for more details about using this specification, please refer to OpenAPI documentations: https://swagger.io/resources/articles/documenting-apis-with-swagger/  # noqa: E501
+    This protocol allows data to be shared with the ELEXIS platform and should be implemented
+    by all providers of data to the ELEXIS platform. This is an OpenAPI documentation,
+    for more details about using this specification,
+    please refer to OpenAPI documentations:
+    https://swagger.io/resources/articles/documenting-apis-with-swagger/  # noqa: E501
 
     The version of the OpenAPI document: 1.0
     Generated by: https://openapi-generator.tech
@@ -12,11 +16,11 @@ import json
 
 import pytest
 
-from edie.model import JsonEntry, Metadata, JsonApiResponse
+from edie.model import Metadata, JsonApiResponse
 from edie.vocabulary import SIZE_OF_DICTIONARY
-from metrics.base import NumberOfSensesEvaluator, PublisherEvaluator, LicenseEvaluator, MetadataQuantityEvaluator, \
-    RecencyEvaluator, ApiMetadataResponseEvaluator, DefinitionOfSenseEvaluator, AvgDefinitionLengthEvaluator, \
-    SizeOfDictionaryEvaluator
+from metrics.base import ApiMetadataResponseEvaluator
+from metrics.metadata import PublisherEvaluator, LicenseEvaluator, SizeOfDictionaryEvaluator, RecencyEvaluator, \
+    MetadataQuantityEvaluator
 
 
 @pytest.fixture(scope="class")
@@ -39,6 +43,7 @@ class TestEntry(unittest.TestCase):
         # FIXME: construct object with mandatory attributes with example values
         # model = Entry()  # noqa: E501
         pass
+
 
 @pytest.mark.usefixtures("metadata_sample")
 class TestPublisherMetadata(unittest.TestCase):
@@ -153,6 +158,7 @@ class TestSizeOfDictionary(unittest.TestCase):
         result = evaluator.result()
 
         self.assertIsNotNone(result[SIZE_OF_DICTIONARY])
+
 
 @pytest.mark.usefixtures("metadata_sample")
 class TestRecency(unittest.TestCase):
